拆分核心功能性能埋点和A/B/C/D埋点
